third_party/blink/web_tests/virtual/css-gap-decorations/external/wpt/css/css-multicol/parsing/column-rule-style-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['column-rule-style'] = "double dashed" should not set the property value
  assert_equals: expected "" but got "double dashed"
Harness: the test ran to completion.
